requirements.txt
fastapi
boto3
numpy
pydantic
sqlalchemy
uvicorn
docx2txt
pypdf
python-dotenv
python-docx
openpyxl
python-pptx
passlib
PyJWT
python-jose
python-multipart
bcrypt
pylint
dependency-check
pipdeptree
pika 
celery
aiokafka 
aiofiles
requests
redis
flower
matplotlib
kafka-python-ng
loguru
PyPDF2
tika
transformers
torch
transformers 
nltk 
scikit-learn
temporalio
six